+++ src/main/java/io/xpring/xrpl/ReliableSubmissionXrpClient.java
@@ -1,6 +1,7 @@
 package io.xpring.xrpl;
 
 import io.xpring.xrpl.model.SendXrpDetails;
+import io.xpring.xrpl.model.TransactionResult;
 import io.xpring.xrpl.model.XrpTransaction;
 
 import java.math.BigInteger;
@@ -63,6 +64,18 @@ public class ReliableSubmissionXrpClient implements XrpClientDecorator {
   @Override
   public XrpTransaction getPayment(String transactionHash) throws XrpException {
     return this.decoratedClient.getPayment(transactionHash);
+  }
+
+  @Override
+  public TransactionResult enableDepositAuth(Wallet wallet) throws XrpException {
+    TransactionResult initialResult = this.decoratedClient.enableDepositAuth(wallet);
+    String transactionHash = initialResult.hash();
+    RawTransactionStatus finalStatus = this.awaitFinalTransactionResult(transactionHash, wallet);
+    return TransactionResult.builder()
+                            .hash(initialResult.hash())
+                            .status(this.getPaymentStatus(transactionHash))
+                            .validated(finalStatus.getValidated())
+                            .build();
   }
 
   private RawTransactionStatus awaitFinalTransactionResult(String transactionHash, Wallet sender) throws XrpException {