+++ src/test/java/io/xpring/xrpl/FakeXrpProtobufs.java
@@ -4,6 +4,7 @@ import com.google.protobuf.ByteString;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.xrpl.rpc.v1.AccountAddress;
+import org.xrpl.rpc.v1.AccountSet;
 import org.xrpl.rpc.v1.CheckCash;
 import org.xrpl.rpc.v1.Common;
 import org.xrpl.rpc.v1.Common.Account;